+++ src/cli.rs
@@ -7,8 +7,8 @@ pub struct Args {
     #[arg(long)]
     pub peer: Option<Multiaddr>,
 
-    #[arg(short, long)]
-    pub listen_address: Option<Multiaddr>,
+    #[arg(long)]
+    pub listen: Option<Multiaddr>,
 
     #[clap(subcommand)]
     pub argument: Argument,
@@ -17,9 +17,6 @@ pub struct Args {
 #[derive(Debug, Parser)]
 pub enum Argument {
     Provide {
-        #[clap(long)]
-        name: String,
-
         #[arg(short, long)]
         wasm: String,
 
@@ -34,44 +31,3 @@ pub enum Argument {
         name: String,
     },
 }
-
-// use cid::multihash::{Code, MultihashDigest};
-// use cid::Cid;
-// use multihash;
-// use std::collections::BTreeMap;
-
-// #[derive(Debug, Clone, Sized)]
-// struct TaskLabel(String);
-//
-// #[derive(Debug, Clone)]
-// enum Input {
-//     WasmParam(Value),
-//     Absolute(Cid),
-//     Relative(TaskLabel)
-// }
-//
-// #[derive(Debug, Sized)]
-// struct Pure {
-//     wasm: Cid,
-//     inputs: [Input]
-// }
-//
-// #[derive(Debug)]
-// struct Effect {
-//     resource: String, // Uri,
-//     action: String
-// }
-//
-// #[derive(Debug)]
-// enum Task {
-//     PureTask(Pure),
-//     EffectTask(Effect)
-// }
-//
-// #[derive(Debug)]
-// struct Job {
-//     // version: semver::Version,
-//     // nonce: String,
-//     tasks: BTreeMap<TaskLabel, Task>
-//     // signature:
-// }
