--- src/ipvm/closure.rs
@@ -1,5 +1,0 @@
-use libipld::{cid::multibase::Base, Ipld, Link};
-use signature::Signature;
-use std::{collections::btree_map::BTreeMap, result::Result};
-use ucan::ucan::Ucan;
-use url::Url;
